calamine
Enum DataRef

pub enum DataRef<'a> {
    Int(i64),
    Float(f64),
    String(String),
    SharedString(&'a str),
    Bool(bool),
    DateTime(ExcelDateTime),
    DateTimeIso(String),
    DurationIso(String),
    Error(CellErrorType),
    Empty,
}

An enum to represent all different data types that can appear as a value in a worksheet cell

Variants

Int(i64)

Signed integer

Float(f64)

Float

String(String)

String

SharedString(&'a str)

Shared String

Bool(bool)

Boolean

DateTime(ExcelDateTime)

Date or Time

DateTimeIso(String)

Date, Time or DateTime in ISO 8601

DurationIso(String)

Duration in ISO 8601

Error(CellErrorType)

Error

Empty

Empty cell