Trait calamine::Reader

pub trait Reader: Sized {
    type RS: Read + Seek;
    type Error: Debug + From<Error>;
    fn new(reader: Self::RS) -> Result<Self, Self::Error>;

    fn vba_project(
        &mut self
    ) -> Option<Result<Cow<'_, VbaProject>, Self::Error>>;

    fn metadata(&self) -> &Metadata;

    fn worksheet_range(
        &mut self, 
        name: &str
    ) -> Option<Result<Range<DataType>, Self::Error>>;

    fn worksheets(&mut self) -> Vec<(String, Range<DataType>)>;

    fn worksheet_formula(
        &mut self, 
        _: &str
    ) -> Option<Result<Range<String>, Self::Error>>;

    fn sheet_names(&self) -> &[String] { ... }

    fn defined_names(&self) -> &[(String, String)] { ... }

    fn worksheet_range_at(
        &mut self, 
        n: usize
    ) -> Option<Result<Range<DataType>, Self::Error>> { ... }
}

A trait to share spreadsheets reader functions accross different FileTypes

Associated Types

type RS: Read + Seek

Inner reader type

type Error: Debug + From<Error>

Error specific to file type

Required methods

fn new(reader: Self::RS) -> Result<Self, Self::Error>

Creates a new instance.

fn vba_project(&mut self) -> Option<Result<Cow<'_, VbaProject>, Self::Error>>

Gets VbaProject

fn metadata(&self) -> &Metadata

Initialize

fn worksheet_range(
&mut self,
name: &str
) -> Option<Result<Range<DataType>, Self::Error>>

Read worksheet data in corresponding worksheet path

fn worksheets(&mut self) -> Vec<(String, Range<DataType>)>

Fetch all worksheet data & paths

fn worksheet_formula(
&mut self,
_: &str
) -> Option<Result<Range<String>, Self::Error>>

Read worksheet formula in corresponding worksheet path

Provided methods

fn sheet_names(&self) -> &[String]

Get all sheet names of this workbook, in workbook order

Examples

use calamine::{Xlsx, open_workbook, Reader};

let mut workbook: Xlsx<_> = open_workbook(path).unwrap();
println!("Sheets: {:#?}", workbook.sheet_names());

fn defined_names(&self) -> &[(String, String)]

Get all defined names (Ranges names etc)

fn worksheet_range_at(
&mut self,
n: usize
) -> Option<Result<Range<DataType>, Self::Error>>

Get the nth worksheet. Shortcut for getting the nth sheet_name, then the corresponding worksheet.