calamine
Trait DataType

pub trait DataType {

    // Required methods
    fn is_empty(&self) -> bool;
    fn is_int(&self) -> bool;
    fn is_float(&self) -> bool;
    fn is_bool(&self) -> bool;
    fn is_string(&self) -> bool;
    fn is_error(&self) -> bool;
    fn get_int(&self) -> Option<i64>;
    fn get_float(&self) -> Option<f64>;
    fn get_bool(&self) -> Option<bool>;
    fn get_string(&self) -> Option<&str>;
    fn get_error(&self) -> Option<&CellErrorType>;
    fn as_string(&self) -> Option<String>;
    fn as_i64(&self) -> Option<i64>;
    fn as_f64(&self) -> Option<f64>;

}

A trait to represent all different data types that can appear as a value in a worksheet cell

Required Methods

fn is_empty(&self) -> bool

Assess if datatype is empty

fn is_int(&self) -> bool

Assess if datatype is a int

fn is_float(&self) -> bool

Assess if datatype is a float

fn is_bool(&self) -> bool

Assess if datatype is a bool

fn is_string(&self) -> bool

Assess if datatype is a string

fn is_error(&self) -> bool

Assess if datatype is a CellErrorType

fn get_int(&self) -> Option<i64>

Try getting int value

fn get_float(&self) -> Option<f64>

Try getting float value

fn get_bool(&self) -> Option<bool>

Try getting bool value

fn get_string(&self) -> Option<&str>

Try getting string value

fn get_error(&self) -> Option<&CellErrorType>

Try getting Error value

fn as_string(&self) -> Option<String>

Try converting data type into a string

fn as_i64(&self) -> Option<i64>

Try converting data type into an int

fn as_f64(&self) -> Option<f64>

Try converting data type into a float