calamine
Trait Reader

pub trait Reader<RS>: Sized
where
    RS: Read + Seek,
{
    type Error: Debug + From<Error>;

    // Required methods
    fn new(reader: RS) -> Result<Self, Self::Error>;
    fn with_header_row(&mut self, header_row: HeaderRow) -> &mut Self;
    fn vba_project(
        &mut self,
    ) -> Option<Result<Cow<'_, VbaProject>, Self::Error>>;
    fn metadata(&self) -> &Metadata;
    fn worksheet_range(
        &mut self,
        name: &str,
    ) -> Result<Range<Data>, Self::Error>;
    fn worksheets(&mut self) -> Vec<(String, Range<Data>)>;
    fn worksheet_formula(
        &mut self,
        _: &str,
    ) -> Result<Range<String>, Self::Error>;

    // Provided methods
    fn sheet_names(&self) -> Vec<String> { ... }
    fn sheets_metadata(&self) -> &[Sheet] { ... }
    fn defined_names(&self) -> &[(String, String)] { ... }
    fn worksheet_range_at(
        &mut self,
        n: usize,
    ) -> Option<Result<Range<Data>, Self::Error>> { ... }
}

A trait to share spreadsheets reader functions across different FileTypes

Required Associated Types

type Error: Debug + From<Error>

Error specific to file type

Required Methods

fn new(reader: RS) -> Result<Self, Self::Error>

Creates a new instance.

fn with_header_row(&mut self, header_row: HeaderRow) -> &mut Self

Set header row (i.e. first row to be read) If header_row is None, the first non-empty row will be used as header row

fn vba_project(&mut self) -> Option<Result<Cow<'_, VbaProject>, Self::Error>>

Gets VbaProject

fn metadata(&self) -> &Metadata

Initialize

fn worksheet_range(&mut self, name: &str) -> Result<Range<Data>, Self::Error>

Read worksheet data in corresponding worksheet path

fn worksheets(&mut self) -> Vec<(String, Range<Data>)>

Fetch all worksheet data & paths

fn worksheet_formula(&mut self, _: &str) -> Result<Range<String>, Self::Error>

Read worksheet formula in corresponding worksheet path

Provided Methods

fn sheet_names(&self) -> Vec<String>

Get all sheet names of this workbook, in workbook order

Examples

use calamine::{Xlsx, open_workbook, Reader};

let mut workbook: Xlsx<_> = open_workbook(path).unwrap();
println!("Sheets: {:#?}", workbook.sheet_names());

fn sheets_metadata(&self) -> &[Sheet]

Fetch all sheets metadata

fn defined_names(&self) -> &[(String, String)]

Get all defined names (Ranges names etc)

fn worksheet_range_at( &mut self, n: usize, ) -> Option<Result<Range<Data>, Self::Error>>

Get the nth worksheet. Shortcut for getting the nth sheet_name, then the corresponding worksheet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.